# BE/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.26.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)


type Product struct {
	ID               int32
	ProfileID        pgtype.Text
	Title            string
	Description      pgtype.Text
	ShortDescription pgtype.Text
	Price            pgtype.Numeric
	Quantity         int32
	DiscountPrice    pgtype.Numeric
	RegularPrice     pgtype.Numeric
	CreatedAt        pgtype.Timestamp
	UpdatedAt        pgtype.Timestamp
	Type             pgtype.Text
}


type ProductExternalBrand struct {
	ID        int32
	Name      string
	LogoUrl   pgtype.Text
	CreatedAt pgtype.Timestamp
	UpdatedAt pgtype.Timestamp
}

type ProductExternalUrl struct {
	ID        int32
	ProductID int32
	BrandID   int32
	ImageUrl  string
	CreatedAt pgtype.Timestamp
	UpdatedAt pgtype.Timestamp
}

type ProductImage struct {
	ID        int32
	ProductID int32
	ImageUrl  string
	CreatedAt pgtype.Timestamp
}

type ProductImageThumbnail struct {
	ProductID      int32
	ProductImageID int32
	CreatedAt      pgtype.Timestamp
}

type ProductTag struct {
	TagID     int32
	ProductID int32
}

type Profile struct {
	ID           string
	CreatedAt    pgtype.Timestamp
	ThumbnailUrl pgtype.Text
	Email        pgtype.Text
	Username     pgtype.Text
}

type ProfileAddress struct {
	ID           int32
	ProfileID    string
	AddressLine1 string
	AddressLine2 pgtype.Text
	PostalCode   string
	Country      string
	City         string
	PhoneNumber  pgtype.Text
}

type ProfileProductCartItem struct {
	ProductID       int32
	ProfileID       string
	ProductQuantity int32
	CreatedAt       pgtype.Timestamp
}

type ProfileProductFavorite struct {
	ProductID int32
	ProfileID string
	CreatedAt pgtype.Timestamp
}

type Tag struct {
	ID        int32
	TagName   string
	CreatedAt pgtype.Timestamp
	UpdatedAt pgtype.Timestamp
}